+++ R/get_model_labels.R
@@ -165,7 +165,13 @@ get_dv_labels <- function(models, case = NULL) {
   # get intercept vectors
 
   intercepts <- purrr::map(models, ~ dplyr::pull(get_model_frame(.x), var = 1))
-  intercepts.names <- purrr::map(models, ~ deparse(stats::formula(.x)[[2L]]))
+  intercepts.names <-
+    purrr::map(models, function(x) {
+      if (inherits(x, "brmsfit"))
+        deparse(stats::formula(x)$formula[[2L]])
+      else
+        deparse(stats::formula(x)[[2L]])
+    })
 
 
   # get all labels